Move sum-to-loop transformation to Fortran intrinsics
