+++ dace/frontend/fortran/ast_transforms.py
@@ -701,23 +701,6 @@ class ArrayLoopNodeLister(NodeVisitor):
         return
 
 
-class SumLoopNodeLister(NodeVisitor):
-    """
-    Finds all sum operations that have to be transformed to loops in the AST
-    """
-    def __init__(self):
-        self.nodes: List[ast_internal_classes.FNode] = []
-
-    def visit_BinOp_Node(self, node: ast_internal_classes.BinOp_Node):
-
-        if isinstance(node.rval, ast_internal_classes.Call_Expr_Node):
-            if node.rval.name.name == "__dace_sum":
-                self.nodes.append(node)
-
-    def visit_Execution_Part_Node(self, node: ast_internal_classes.Execution_Part_Node):
-        return
-
-
 def par_Decl_Range_Finder(node: ast_internal_classes.Array_Subscript_Node,
                           ranges: list,
                           rangepos: list,
@@ -891,114 +874,6 @@ def mywalk(node):
         node = todo.popleft()
         todo.extend(iter_child_nodes(node))
         yield node
-
-
-class SumToLoop(NodeTransformer):
-    """
-    Transforms the AST by removing array sums and replacing them with loops
-    """
-    def __init__(self, ast):
-        self.count = 0
-        ParentScopeAssigner().visit(ast)
-        self.scope_vars = ScopeVarsDeclarations()
-        self.scope_vars.visit(ast)
-
-    def visit_Execution_Part_Node(self, node: ast_internal_classes.Execution_Part_Node):
-        newbody = []
-        for child in node.execution:
-            lister = SumLoopNodeLister()
-            lister.visit(child)
-            res = lister.nodes
-            if res is not None and len(res) > 0:
-
-                current = child.lval
-                val = child.rval
-
-                rvals = []
-                for i in mywalk(val):
-                    if isinstance(i, ast_internal_classes.Call_Expr_Node) and i.name.name == '__dace_sum':
-
-                        for arg in i.args:
-
-                            # supports syntax SUM(arr)
-                            if isinstance(arg, ast_internal_classes.Name_Node):
-                                array_node = ast_internal_classes.Array_Subscript_Node(parent=arg.parent)
-                                array_node.name = arg
-
-                                # If we access SUM(arr) where arr has many dimensions,
-                                # We need to create a ParDecl_Node for each dimension
-                                dims = len(self.scope_vars.get_var(node.parent, arg.name).sizes)
-                                array_node.indices = [ast_internal_classes.ParDecl_Node(type='ALL')] * dims
-
-                                rvals.append(array_node)
-
-                            # supports syntax SUM(arr(:))
-                            if isinstance(arg, ast_internal_classes.Array_Subscript_Node):
-                                rvals.append(arg)
-
-                if len(rvals) != 1:
-                    raise NotImplementedError("Only one array can be summed")
-                val = rvals[0]
-                rangeposrval = []
-                rangesrval = []
-
-                par_Decl_Range_Finder(val, rangesrval, rangeposrval, self.count, newbody, self.scope_vars, True)
-
-                # Initialize the result variable
-                newbody.append(
-                    ast_internal_classes.BinOp_Node(
-                        lval=current,
-                        op="=",
-                        rval=ast_internal_classes.Int_Literal_Node(value="0"),
-                        line_number=child.line_number
-                    )
-                )
-                range_index = 0
-                body = ast_internal_classes.BinOp_Node(lval=current,
-                                                       op="=",
-                                                       rval=ast_internal_classes.BinOp_Node(
-                                                           lval=current,
-                                                           op="+",
-                                                           rval=val,
-                                                           line_number=child.line_number),
-                                                       line_number=child.line_number)
-                for i in rangesrval:
-                    initrange = i[0]
-                    finalrange = i[1]
-                    init = ast_internal_classes.BinOp_Node(
-                        lval=ast_internal_classes.Name_Node(name="tmp_parfor_" + str(self.count + range_index)),
-                        op="=",
-                        rval=initrange,
-                        line_number=child.line_number)
-                    cond = ast_internal_classes.BinOp_Node(
-                        lval=ast_internal_classes.Name_Node(name="tmp_parfor_" + str(self.count + range_index)),
-                        op="<=",
-                        rval=finalrange,
-                        line_number=child.line_number)
-                    iter = ast_internal_classes.BinOp_Node(
-                        lval=ast_internal_classes.Name_Node(name="tmp_parfor_" + str(self.count + range_index)),
-                        op="=",
-                        rval=ast_internal_classes.BinOp_Node(
-                            lval=ast_internal_classes.Name_Node(name="tmp_parfor_" + str(self.count + range_index)),
-                            op="+",
-                            rval=ast_internal_classes.Int_Literal_Node(value="1")),
-                        line_number=child.line_number)
-                    current_for = ast_internal_classes.Map_Stmt_Node(
-                        init=init,
-                        cond=cond,
-                        iter=iter,
-                        body=ast_internal_classes.Execution_Part_Node(execution=[body]),
-                        line_number=child.line_number)
-                    body = current_for
-                    range_index += 1
-
-                newbody.append(body)
-
-                self.count = self.count + range_index
-            else:
-                newbody.append(self.visit(child))
-        return ast_internal_classes.Execution_Part_Node(execution=newbody)
-
 
 class RenameVar(NodeTransformer):
     def __init__(self, oldname: str, newname: str):
